[refactor] Extract Interface  - SOLID 원칙을 위해서 인터페이스 추출
 related to #7
